+++ jscomp/build_tests/white space/yy.ml
@@ -1,6 +1,5 @@
 type t = 
   | Foo
-  [@@bs.deriving jsConverter]
 
 let u = Xx.sum 3